to merge

+++ src/Backend/src/BackExternal/ModelManager.java
@@ -126,7 +126,7 @@ public class ModelManager implements IModelManager {
 
     @Override
     public int getCurrentActiveTurtle() {
-        return myTurtleModel.getCurrentTurtleIndex();
+        return (int) myTurtleModel.getCurrentTurtleIndex();
     }
 
     @Override
